Add install modules
